Integration between branches
Added a tag to contestant feature file since the test written in it is
testing something that is handled somewhere else.
Commented out the code written in competitorSteps steps since the
competitorScore class is not being used at the moment and has old code.

+++ src/main/java/sprint1/ContestantSportEvent.java
@@ -2,10 +2,10 @@ package sprint1;
 
 public class ContestantSportEvent {
 	private String sportEvent;
-	private Double score;
+	private int score;
 	private Double result;
 	
-	public ContestantSportEvent(String sportEvent, Double result, Double score) {
+	public ContestantSportEvent(String sportEvent, Double result, int score) {
 		this.sportEvent = sportEvent;
 		this.score = score;
 		this.result = result;
@@ -19,7 +19,7 @@ public class ContestantSportEvent {
 		return sportEvent;
 	}
 
-	public Double getScore() {
+	public int getScore() {
 		return score;
 	}
